Initial project skeletons
This commit adds the skeleton for Spring IO Platform 1.0. It provides
a pom that can be imported or used as a parent. In subsequent
commits, this pom will be updated to provide dependency management for
the Platform's components and their dependencies. It also provides a
properties file that contains groupId:artifactId=version pairs. To
ensure consistency, this properties file is derived from the pom using
an XSLT. The intention is that Boot's Gradle plugin will be updated to
allow it to be configured to use these versions as an alternative to
those provided by default.