Add third-party dependencies not covered by Boot
Add all of the third-party dependencies for which Spring Boot does not
already provide dependency management.

Spring LDAP includes an exclusion to remap the group id of a transitive
dependency on Google Guice from openid4java-nodeps as the coordinates
that are used do not exist in Maven Central. The verifyDependencies
script has been updated to cope with this.

A new script, findDuplicates, has been added. This script can be used
to compare the Platform's pom with Spring Boot's and report any
duplicates.