Only display the API doc link if the project has API docs
This is something of a hack at the moment. The right border on the
reference doc link has been removed. This border provides a separator
between the reference doc link and the API doc link. It’s fine for the
platform as it has no API docs but it’s not a general solution.