Rename modules: platform -> platform-bom and platform-docs -> platform
Rename the platform artifact to platform-bom to better describe what
it is; it's the platform's bill of materials. Also renamed
platform-docs to platform. This is the artifact that represents the
Platform as a whole. All it does at the moment is publish the
Platform's docs but this may expand in the future.