Use the latest available snapshots of projects that have yet to release
Projects that have yet to release their Spring IO Platform version have
been updated to use snapshots so that the Platform provides the version
that is as close as possible to what will be released.

As part of this Boot has been updated to 1.1.0.BUILD-SNAPSHOT. As a
result of this the Platform’s dependency management has been updated to
provide complete coverage while also avoiding duplicating what’s
provided by Boot.