Build on top of the new separate Platform Foundation bom
This commit updates the main Spring IO Platform to split out the
Platform's Foundation and start building on top of it.

This change does not affect existing users of the Platform as the
Platform's bom and the Platform's versions properties file are both
still published with the same dependency management at the same Maven
coordinates. However, the properties file published by
platform-versions should now be considered deprecated and the
properties file published by platform-bom should be used instead.
The intention is to stop publishing the platform-versions properties
artifact in the 2.0 release of the Platform.