Improve automated testing of the Platform
Prior to this commit, analysis of the Platform was only performed
manually via the analyzers in the Platform Tools project. This commit
moves some of that analysis into the Platform project in the form of
tests driven via JUnit. The Platform is now automatically analyzed
for missing dependencies (dependencies that are used by something in
the Platform but are not included in the Platform's dependency
management) and unused dependencies (dependencies that are included in
the Platform's dependency management but are not used by anything in
the Platform).