Update the Platform definition to match latest Boot snapshots
The latest Boot snapshots have updated to Spring Data Dijksta SR 4.
This commit updates the versions of the Platform's Spring Data
projects accordingly.

The latest Boot snapshots have added dependency management for
ActiveMQ's JMS Pool. This dependency is not used by Boot or any of
the other projects in the Platform. This commit marks it as being
included for user convenience.