Don't use the Foundation in 1.1.x as XD will not be included in 1.1
The Platform Foundation bom was introduced to allow Spring XD to
become part of the Platform while also using the Platform for its
dependency management. Without the foundation this would have created
a circular dependency. The foundation was intended to provide XD with
its dependency management will allowing the main Platform bom to
depend on XD. XD is no longer going to be part of Spring IO Platform
1.1 so the use of the foundation is adding unnecessary complexity.

This commit removes the use of the Foundation from the 1.1.x branch.
Its usage will be retained in master, for the time being at least.