Workaround Artifactory promotion limitation
Artifactory will not promote two artifacts with the same hash, even
if the two artifacts have different coordinates. This causes a
problem as the platform-bom.properties and
platform-versions.properties files have identical contents (this
is intentional with platform-versions.properties being deprecated in
favour of the new platform-bom.properties file).

This commit works around the problem by appending a new line to the
end of platform-versions.properties thereby ensuring that it has
a different hash to platform-bom.properties while still being
functionaly identical.

Closes #122