Use a prefix, default ‘v’, to map versions to tags
Previously, when a project deviated from the convention of prefixing
a version with ‘v’ in its tags, the whole tag had to be specified as a
checkout attribute. This meant that the changes had to be made in two
places when the version of such a project was updated. This commit
reworks the logic to use a prefix instead. This means that a change to
a project’s version now only requires a single change in the definition.