Upgrade Kryo to 3.0.3
The Spring XD has moved its Kryo Codecs to SI Core,
hence we have there a new Kryo dependency.

According to the README on the Kryo site (https://github.com/EsotericSoftware/kryo)
it is better to use `kryo-shaded` artifact to avoid conflicts with our ASM jar.

So, this fix change `kryo` artifact to `kryo-shaded` and raise its version to the `3.0.3`