Add patch to ensure that SI Java DSL tests have an MBeanServer bean
Some tests in the Spring Integration Java DSL were relying on some
faulty logic in Boot's IntegrationAutoConfiguration that meant that
an MBeanServer bean was auto-configured even when JMX
auto-configuration was disabled. This was fixed in
spring-projects/spring-boot#5309. This patch updates the tests so that
they pass when run with versions of Spring Boot that correct the
faulty logic.