GH-341: Downgrade to curator-2.10.0. Remove ZK
Fixes GH-341 (https://github.com/spring-io/platform/issues/341)

The previous upgrade to `curator-3.1.0` was wrong, because it isn't compatible with the latest stable ZK (3.4.8)
And pulls `alpha` (3.5.1) version as transitive dependency.
That makes our test incompatible when platform re-map transitive version to our stable.

* Downgrade to `curator-2.10.0` according their note (http://curator.apache.org/):

> The are currently two released versions of Curator, 2.x.x and 3.x.x:

>    Curator 2.x.x - compatible with both ZooKeeper 3.4.x and ZooKeeper 3.5.x
>    Curator 3.x.x - compatible only with ZooKeeper 3.5.x and includes support for new features such as dynamic reconfiguration, etc.

To avoid the `alpha` version in the IO.

* Remove an explicit `ZK` dependency in favor of the transitive from the `curator`.