Use a dash separator (rather than a dot) in the version
A dot separator has the problem of making the version resemble that of
a semantic version where something may attempt to parse the version by
splitting on dot characters. Using a dash helps to avoid this
misinterpretation of the version. This also aligns the Platform’s
version scheme with that of Spring Data.

Closes gh-407